Fixed capitalization in file name when dbmanager.py gets created.
Added a type check to convert string to list[string]
when "columns_to_return" is passed to get_rows() as a string
and "columns" is passed to find() as a string.
Prevents me from going insane when I only need to pass one column
and forget to wrap it in braces.